// src/generated/models/Dnx_projectsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Dnx_projectsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Dnx_projectsstatecode = keyof typeof Dnx_projectsstatecode;
export const Dnx_projectsstatuscode = {
  1: 'Active',
  2: 'Inactive',
  778540001: 'Archived',
  778540002: 'OnHold'
} as const;
export type Dnx_projectsstatuscode = keyof typeof Dnx_projectsstatuscode;

export interface Dnx_projectsBase {
  dnx_description?: string;
  dnx_project_code?: string;
  dnx_project_name: string;
  dnx_projectid: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Dnx_projectsstatecode;
  statuscode?: Dnx_projectsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Dnx_projects extends Dnx_projectsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}